Refactor normalization service infrastructure
