+++ src/bioetl/domain/transform/impl/__init__.py
@@ -3,11 +3,15 @@ Transform implementations.
 """
 
 from .hasher import HasherImpl
-from .normalize import NormalizationService, serialize_dict, serialize_list
+from .normalize import normalize_pubchem_cid, normalize_pubmed_id, normalize_scalar, normalize_uniprot_id
+from .serializer import serialize_dict, serialize_list
 
 __all__ = [
     "HasherImpl",
-    "NormalizationService",
+    "normalize_scalar",
+    "normalize_pubmed_id",
+    "normalize_pubchem_cid",
+    "normalize_uniprot_id",
     "serialize_dict",
     "serialize_list",
 ]
