+++ src/bioetl/infrastructure/clients/base/abc_impls.yaml
@@ -32,6 +32,11 @@ ValidatorABC:
   implementations:
     Pandera: bioetl.domain.validation.impl.pandera_validator.PanderaValidatorImpl
 
+NormalizationServiceABC:
+  default_factory: bioetl.infrastructure.transform.factories.default_normalization_service
+  implementations:
+    Default: bioetl.infrastructure.transform.impl.normalization_service_impl.NormalizationServiceImpl
+
 WriterABC:
   default_factory: bioetl.infrastructure.output.factories.default_writer
   implementations:
