+++ src/bioetl/infrastructure/transform/__init__.py
@@ -1,0 +1,5 @@
+"""Transform infrastructure layer."""
+
+from bioetl.infrastructure.transform.factories import default_normalization_service
+
+__all__ = ["default_normalization_service"]
