+++ src/bioetl/infrastructure/transform/impl/__init__.py
@@ -1,0 +1,7 @@
+"""Concrete transform implementations."""
+
+from bioetl.infrastructure.transform.impl.normalization_service_impl import (
+    NormalizationServiceImpl,
+)
+
+__all__ = ["NormalizationServiceImpl"]
